pyproject.toml
[project]
name = "kokoro-onnx"
version = "0.4.9"
description = "TTS with kokoro and onnx runtime"
readme = "README.md"
authors = [
    { name = "thewh1teagle", email = "61390950+thewh1teagle@users.noreply.github.com" },
]
requires-python = ">=3.10,<3.14"
dependencies = [
    "onnxruntime>=1.20.1",
    "colorlog>=6.9.0",
    "espeakng-loader>=0.2.4",
    "phonemizer-fork>=3.3.2",
    "numpy>=2.0.2",
    "soundfile>=0.13.0",
    "pypdf2>=3.0.1",
    "ffmpeg-python>=0.2.0",
]

[project.urls]
Homepage = "https://github.com/thewh1teagle/kokoro-onnx"
Repository = "https://github.com/thewh1teagle/kokoro-onnx"
Issues = "https://github.com/thewh1teagle/kokoro-onnx/issues"


[project.optional-dependencies]
# Windows/Linux GPU feature
# Install with kokoro-onnx[gpu]
gpu = [
    # onnxruntime-gpu is not available on Linux ARM or macOS
    "onnxruntime-gpu>=1.20.1; platform_machine == 'x86_64' and sys_platform != 'darwin'",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = ["ruff>=0.11.0", "sounddevice>=0.5.1", "soundfile>=0.13.0"]

[tool.ruff]
required-version = ">=0.9.0"
output-format = "concise"
show-fixes = true

[tool.ruff.lint]
extend-select = ["I", "UP"]

[tool.ruff.lint.isort]
split-on-trailing-comma = false
